Move `ValidateMetadata` in `validator/helpers`.
