+++ validator/slashing-protection-history/import.go
@@ -60,7 +60,7 @@ func importStandardProtectionJSONComplete(ctx context.Context, validatorDB *kv.S
 	}
 
 	// We validate the `MetadataV0` field of the slashing protection JSON file.
-	if err := validateMetadata(ctx, validatorDB, interchangeJSON); err != nil {
+	if err := helpers.ValidateMetadata(ctx, validatorDB, interchangeJSON); err != nil {
 		return errors.Wrap(err, "slashing protection JSON metadata was incorrect")
 	}
 
@@ -156,7 +156,7 @@ func importStandardProtectionJSONMinimal(ctx context.Context, validatorDB db.Dat
 	}
 
 	// We validate the `MetadataV0` field of the slashing protection JSON file.
-	if err := validateMetadata(ctx, validatorDB, interchangeJSON); err != nil {
+	if err := helpers.ValidateMetadata(ctx, validatorDB, interchangeJSON); err != nil {
 		return errors.Wrap(err, "slashing protection JSON metadata was incorrect")
 	}
 
@@ -317,40 +317,6 @@ func initializeProgressBar(numItems int, msg string) *progressbar.ProgressBar {
 		progressbar.OptionOnCompletion(func() { fmt.Println() }),
 		progressbar.OptionSetDescription(msg),
 	)
-}
-
-func validateMetadata(ctx context.Context, validatorDB db.Database, interchangeJSON *format.EIPSlashingProtectionFormat) error {
-	// We need to ensure the version in the metadata field matches the one we support.
-	version := interchangeJSON.Metadata.InterchangeFormatVersion
-	if version != format.InterchangeFormatVersion {
-		return fmt.Errorf(
-			"slashing protection JSON version '%s' is not supported, wanted '%s'",
-			version,
-			format.InterchangeFormatVersion,
-		)
-	}
-
-	// We need to verify the genesis validators root matches that of our chain data, otherwise
-	// the imported slashing protection JSON was created on a different chain.
-	gvr, err := helpers.RootFromHex(interchangeJSON.Metadata.GenesisValidatorsRoot)
-	if err != nil {
-		return fmt.Errorf("%#x is not a valid root: %w", interchangeJSON.Metadata.GenesisValidatorsRoot, err)
-	}
-	dbGvr, err := validatorDB.GenesisValidatorsRoot(ctx)
-	if err != nil {
-		return errors.Wrap(err, "could not retrieve genesis validators root to db")
-	}
-	if dbGvr == nil {
-		if err = validatorDB.SaveGenesisValidatorsRoot(ctx, gvr[:]); err != nil {
-			return errors.Wrap(err, "could not save genesis validators root to db")
-		}
-		return nil
-	}
-	if !bytes.Equal(dbGvr, gvr[:]) {
-		return errors.New("genesis validators root doesn't match the one that is stored in slashing protection db. " +
-			"Please make sure you import the protection data that is relevant to the chain you are on")
-	}
-	return nil
 }
 
 // We create a map of pubKey -> []*SignedBlock. Then, for each public key we observe,
